requirements.txt
mcp>=0.9.0
python-dotenv>=1.0.0
Pillow>=10.0.0
requests>=2.31.0
flask>=2.0.0
pycryptodome>=3.18.0
cryptography>=38.0.0
sqlalchemy>=2.0.0